feature(navbar): Display enabled, approved count next to org #342
This feature adds a slice of what is requested in #342, namely the
organization counts. These can be found as badges next to organizations
in the Orgranizations pull down. The data comes from calling an existing
grails backend endpoint (which had to be modified slightly to allow
users - not just org stewards and admin - to access the endpoint). We
also had to fork the react-select-box repo to allow adding tags with
count information to options.

In terms of React/Reflux, the Organizations component listens to a new
store, CountsStore which in turn listens to a new action,
fetchCounts. fetchCounts calls the ListingApi with the enabled and
approved query parameters set.